requirements.txt
flask==3.1.0
tensorflow-cpu==2.19.0
numpy==2.1.3
pillow==11.2.1
gunicorn==23.0.0